+++ examples/test_runner/config/order_config.yml
@@ -2,5 +2,5 @@
 host: http://47.103.74.144
 #host: http://192.168.1.23
 symbol: XBTUSD
-api_key: x45yReH6K0OEHiyaYh59
-api_secret: 1QbO799k7A9fDce44E86w4JP9AL64N3HUi6Cwjg7v1mrvz16QP0qTbKiqX9Nlb7fJlf0C9zp0EXo8R103KZHJVgGzbtfn42MSW8
+api_key: 1B63T5cP9xjc4QVxuwDD
+api_secret: 52QT8279xmFXtKK0X2SMKkcf07yv6e2zb79PGAhy1Hu8Fh1FiQgnK6G1Q2c6aEug936JH536yyLaLAz9b3HcNvj4GYTD5sM02xr
